# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "description": "This is an example API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Login an Admin account",
                "description": "Returns admin data",
                "operationId": "02d2894f743c006b2a0ce550fa19a903",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "List categories",
                "description": "Returns categories",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page"
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort By"
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Desc",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create new Category Item",
                "description": "Create new Category Item",
                "operationId": "dd76b8d73b7ea8b4951f03d7c0904c92",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/categories/{uuid}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Show Category by uuid",
                "description": "Returns categories",
                "operationId": "8c905f3c21efee2dce9b6cea3af9b23e",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Update Category Item",
                "description": "Update Category Item",
                "operationId": "5eb2b6744da5693d2c3fcd357e76d6d8",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Delete Category by uuid",
                "description": "Delete a Category",
                "operationId": "e3ca4d164f0506d88aeffa50eb3b98ca",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "List products",
                "description": "Returns Products",
                "operationId": "1ce27f90dd8c39cea9520bf45d58c7c2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Per page"
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort By"
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Desc",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create new Product Item",
                "description": "Create new Product Item",
                "operationId": "acaa0d3fdc74fd4b534e70fc016cf7f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/products/{uuid}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Show Product by uuid",
                "description": "Returns product",
                "operationId": "a6661e787c8855e2e958979cde18b125",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product Item",
                "description": "Update product Item",
                "operationId": "1feb43f2cf58693ece0a63b45635ceb7",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete Product by uuid",
                "description": "Delete a Product",
                "operationId": "fe3fbe51a4ab866c26b88d94463c76ac",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "uuid",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login an User account",
                "description": "Returns User's Token",
                "operationId": "2466229c3e831e8f1e633382db34ffe3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login user",
                "description": "Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email of the user",
                        "type": "string",
                        "format": "email",
                        "example": "user@mail.com"
                    },
                    "password": {
                        "description": "Password for user's account",
                        "type": "string",
                        "example": "mypassword"
                    }
                },
                "type": "object"
            },
            "CategoryStoreRequest": {
                "title": "Category Store Request",
                "description": "Category Store Request body data",
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "description": "Name/title of Category for a product",
                        "type": "string",
                        "example": "dry dog food"
                    }
                },
                "type": "object"
            },
            "ProductStoreRequest": {
                "title": "Product Store Request",
                "description": "Product Store Request body data",
                "required": [
                    "title",
                    "category_uuid",
                    "price",
                    "description"
                ],
                "properties": {
                    "title": {
                        "description": "Name/title of Product",
                        "type": "string",
                        "example": "dry dog food"
                    },
                    "category_uuid": {
                        "description": "Category for this product",
                        "type": "string"
                    },
                    "price": {
                        "description": "Price of Product",
                        "type": "number",
                        "format": "float",
                        "example": "100"
                    },
                    "description": {
                        "description": "Description of Product",
                        "type": "string",
                        "example": "This food is absolutely for your dogs"
                    }
                },
                "type": "object"
            }
        }
    }
}